[WFCORE-5442] Math operands should be cast before assignment

+++ deployment-scanner/src/main/java/org/jboss/as/server/deployment/scanner/ZipCompletionScanner.java
@@ -283,7 +283,7 @@ public class ZipCompletionScanner {
                 break;
             }
             lastChannelPos = channelPos;
-            channelPos -= Math.min(channelPos - bufferPos, CHUNK_SIZE - bufferPos);
+            channelPos -= Math.min(channelPos - bufferPos, (long) CHUNK_SIZE - bufferPos);
         }
 
         return false;
